requirements.txt

amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.7
cx-Oracle==8.3.0
Django==4.2.7
django-rangefilter==0.9.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
ecdsa==0.18.0
et-xmlfile==1.1.0
gunicorn==21.2.0
idna==3.6
jose==1.0.0
kombu==5.3.5
msal==1.25.0
numpy==1.26.4
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
prompt-toolkit==3.0.43
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-jose==3.3.0
pytz==2023.3.post1
redis==5.0.3
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
django-background-tasks
xmltodict
djangorestframework-xml